backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
python-dotenv==1.2.1
pandas>=2.0.0
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
